requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
slowapi==0.1.9
json-log-formatter==0.5.2

SQLAlchemy==2.0.34
psycopg2-binary==2.9.9

chromadb>=1.0.10

sentence-transformers==2.7.0
transformers>=4.41,<5

httpx==0.27.2
pandas==2.2.2
openpyxl==3.1.5
beautifulsoup4==4.12.3
watchdog==6.0.0
llama-cpp-python==0.3.16

openai==2.7.2

